Factor out logic transforming participant Redux representations into daily-js-consumer-facing representations into its own file, Participant
